Move persistence tests into persist module
